👌 IMPROVE: Add fix for LazyVim autocmd 'last_loc'
When exiting vim without saving the current buffer, the last location
might be in a line number that's greater thatn the number of lines
in the buffer. In this case, the 'last_loc' autocmd provided by LazyVim
will not set the location and the cursor will be placed at the first
line. I find it more intuitive to place the cursor at the last line in
this case.