🐛 FIX: nvim: Disable mason for lsps only when they are configured
Setting the `mason` property on an lsp server also configures it for
lsp-config. This led to the situation that `rust_analyzer` was
configured even though LazyVim switched to `rustaceanvim`. I suspect
this is the cause for the bug I had where two instances of rust_analyzer
were active at the same time resulting in duplicate inlay hints,
diagnostics and code actions.